requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
PyMySQL
passlib[bcrypt]
python-jose[cryptography]
httpx
fastapi-cors